Trigger top level type inference resolution by member lookup.
Every time we look up a member during type level type inference, we do
so because we need the type of the looked up member.  So rather than
look it up and then trigger its type to resolve, we simply resolve it
as part of the lookup.

Change-Id: I0428649fbc99966dae1c0c3a04b6c5e4f680c9ab
Reviewed-on: https://dart-review.googlesource.com/12840
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>