Don't try to resolve/finalize dependent forwarding stubs during "prepare" phase.
Resolution of method forwarding nodes, and inference of method types,
happens during the "prepare" phase of type inference, since in theory
method types and method forwarding nodes can only depend on methods in
base classes.  Resolution of getter/setter forwarding nodes, and
inference of getter/setter/field types, happens later, after all
forwarding nodes have been created.  This works properly for code that
is free of errors.

But if the code has errors, it's possible that a method might try to
override a getter, setter, or field.  If this happens, we don't want
resolution of the method forwarding node, or inference of the method,
to trigger premature type inference of the getter/setter/field; if it
did, this might cause us to infer the wrong type for the
getter/setter/field, leading to user confusion.

This CL adds a check to ensure that we don't resolve any
getter/setter/field nodes during the "prepare" phase of type
inference.

Change-Id: I794d6ba5fde7937bbec3cf7c21a1c3c602e6bbd1
Reviewed-on: https://dart-review.googlesource.com/14364
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>