Remove unnecessary toValue conversions in conditional handling.
When compiling an if statement, the condition is converted to a value
via popForValue() in beginThenStatement(), so when the condition is
later popped in endIfStatement(), we can just pop() it directly.

A similar situation exists for the "condition" and "then" parts of a
conditional expression, and for the "receiver" part of a logical
operation.

Change-Id: I9f9f13263b284c99fbbfeb38e4a987d0d4fe9e30
Reviewed-on: https://dart-review.googlesource.com/15380
Reviewed-by: Peter von der Ahé <ahe@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>