Refactoring failures tool
Refactoring of failures tool with:

Reducing the time it takes for looking up status files by dividing entries in
the status file by wildcard and none-wildcard. Speciel attention is needed for
multitests.

Not report an error when Skip or SkipByDesign is added to the status files.

Report which status files expectations are taken from.

Bug:
Change-Id: I56a05a9bd79a5206ef5287a58d341165d98aaefc
Reviewed-on: https://dart-review.googlesource.com/16441
Reviewed-by: William Hesse <whesse@google.com>