dart2js kernel: make JClosureField implement PrivatelyNamedJSEntity for nicer names
This change reduces the number of diffs between AST and Kernel by a factor of 3.

I changed the order of adding closure fields - adding fields for
boxes, then 'this', then values and types.  This makes the order more
consistent with the old backend. We can undo this later.

Change-Id: I54d7c601cae89b6debce5166045dbb4d54add91f
Reviewed-on: https://dart-review.googlesource.com/18160
Commit-Queue: Stephen Adams <sra@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Emily Fortuna <efortuna@google.com>