[vm, gc] Handle image pages in the sliding compactor via range checks.
This allows running the sliding compactor for AOT and App-JIT.

The loaders on Mac and Linux do not respect alignments greater than the virtual page size, and we must use the system loader on iOS to get the executable image page. In the future, we might try to organize instructions for AOT/App-JIT off of an image page.

dart2js hello compaction time (without marking):
30249 us -> 31351 us (+3.6%) N=10

Bug: https://github.com/dart-lang/sdk/issues/30978
Change-Id: I6e9ffd04df30ef0a09f1b123ae0c148f11448627
Reviewed-on: https://dart-review.googlesource.com/19140
Reviewed-by: Erik Corry <erikcorry@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>