Update precompiled test status file
A test introduced in https://dart-review.googlesource.com/c/sdk/+/21080
still did not have the correct expectations for the VM precompiler.

The test was marked as crashing except on arm and simarm64, but it does not
crash on x64 except in debug mode.  (I'm not actually sure that it crashes
in all the other configurations that we say it crashes in, but it definitely
does not crash in x64 release or product modes.)

Bug:
Change-Id: I25a8f3cf6cb91cf128cd715d1aaeca2e87d26be4
Reviewed-on: https://dart-review.googlesource.com/22080
Reviewed-by: Erik Ernst <eernst@google.com>
Commit-Queue: Kevin Millikin <kmillikin@google.com>