Apply resolution for BinaryExpression.
I changed the offset the resulting type for invocations. If there are
arguments (usually there are) we record with the offset of `(`. If there
are no arguments (binary operator), we record with the fileOffset of
the invocation, which is then the offset of the operator.

R=brianwilkerson@google.com, paulberry@google.com

Bug:
Change-Id: Ib5b531440e0782a00764ad68c20e2f1d52ff92d0
Reviewed-on: https://dart-review.googlesource.com/24020
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>